oracle rest bug fix and align syntax
